Split pipeline.rs and render_native.rs at clean top-level boundaries to clear max-lines
