+++ crates/almide-mir/src/pipeline.rs
@@ -714,1287 +714,5 @@ fn build_ir_with_drops(
     Ok(ir)
 }
 
-/// The rest of the pipeline after [`build_ir_with_drops`]: collect globals/layouts,
-/// lower every fn to MIR (main + linked module siblings), synthesize the global-init
-/// and self-host-runtime auto-link fns, then render the final wasm module.
-/// The four lookup tables [`collect_pipeline_layouts`] builds: globals (both the shared
-/// program+modules union and the MAIN-region-bridged view), and the record/variant
-/// layout registries — everything the fn-lowering calls below consult by reference.
-struct PipelineLayouts {
-    globals: HashMap<almide_ir::VarId, almide_lang::types::Ty>,
-    global_inits: HashMap<almide_ir::VarId, almide_ir::IrExpr>,
-    main_globals: HashMap<almide_ir::VarId, almide_lang::types::Ty>,
-    main_global_inits: HashMap<almide_ir::VarId, almide_ir::IrExpr>,
-    mutable_toplet_aliases: std::collections::HashMap<almide_ir::VarId, almide_ir::VarId>,
-    record_layouts: crate::lower::RecordLayouts,
-    variant_layouts: crate::lower::VariantLayouts,
-}
-
-/// Phase 2: collect top-level `let` globals (VarId -> Ty) + their INITIALIZER exprs
-/// (union of program + modules), bridge the MAIN-region view across cross-module
-/// references, and build the record/variant layout registries (aliasing each
-/// UNIQUELY-owned base name onto its qualified layout).
-fn collect_pipeline_layouts(ir: &almide_ir::IrProgram) -> PipelineLayouts {
-    let mut globals: HashMap<almide_ir::VarId, almide_lang::types::Ty> = HashMap::new();
-    let mut global_inits: HashMap<almide_ir::VarId, almide_ir::IrExpr> = HashMap::new();
-    // An UNANNOTATED option-ctor top-let (`let MAYBE = some(Cfg { .. })`) leaves
-    // tl.ty Unknown(-payload) — refine it from the ctor's payload type so the
-    // reference site materializes a REAL tracked Option (see
-    // `refine_option_toplet_ty`; the same repair the crossmod bridge applies).
-    let toplet_ty = |tl: &almide_ir::IrTopLet| {
-        crate::lower::refine_option_toplet_ty(&tl.ty, &tl.value).unwrap_or_else(|| tl.ty.clone())
-    };
-    for tl in &ir.top_lets {
-        globals.insert(tl.var, toplet_ty(tl));
-        global_inits.insert(tl.var, tl.value.clone());
-    }
-    for m in &ir.modules {
-        for tl in &m.top_lets {
-            globals.insert(tl.var, toplet_ty(tl));
-            global_inits.insert(tl.var, tl.value.clone());
-        }
-    }
-    // PER-REGION globals: the shared union above keys BOTH the main program's and each
-    // module's top-let VarIds — two PRIVATE numbering regions that can COLLIDE (main-side
-    // VarId(2) vs a module's VarId(2) are unrelated). MAIN functions must resolve through
-    // main's own entries first (re-inserted last, winning collisions) plus the cross-module
-    // NAME bridge (`toplib.SYSTEM` referenced through a main-side id); MODULE functions keep
-    // the module-entries-win union (their region, as today).
-    let mut main_globals = globals.clone();
-    let mut main_global_inits = global_inits.clone();
-    let mut mutable_toplet_aliases: std::collections::HashMap<almide_ir::VarId, almide_ir::VarId> =
-        std::collections::HashMap::new();
-    crate::lower::bridge_cross_module_toplets(ir, &mut main_globals, &mut main_global_inits, &mut mutable_toplet_aliases);
-    for tl in &ir.top_lets {
-        main_globals.insert(tl.var, toplet_ty(tl));
-        main_global_inits.insert(tl.var, tl.value.clone());
-    }
-
-    // Record-layout registry (type name → fields) for the VALUE MODEL.
-    let mut record_layouts = crate::lower::build_record_layouts(&ir.type_decls);
-    for m in &ir.modules {
-        record_layouts.extend(crate::lower::build_record_layouts(&m.type_decls));
-    }
-    // Alias each UNIQUELY-owned base name to its qualified layout (a bare `Named` reference to a
-    // module record must resolve its field layout); an ambiguous base stays qualified-only.
-    {
-        let mut owners: std::collections::HashMap<String, Vec<String>> = Default::default();
-        for k in record_layouts.keys() {
-            if let Some((_, base)) = k.rsplit_once('.') {
-                owners.entry(base.to_string()).or_default().push(k.clone());
-            }
-        }
-        for (base, ks) in owners {
-            if ks.len() == 1 && !record_layouts.contains_key(&base) {
-                let v = record_layouts.get(&ks[0]).cloned().unwrap();
-                record_layouts.insert(base, v);
-            }
-        }
-    }
-
-    // Variant-layout registry (type name → tag + per-constructor fields) for custom ADTs.
-    let mut variant_layouts = crate::lower::build_variant_layouts(&ir.type_decls);
-    for m in &ir.modules {
-        let m_vl = crate::lower::build_variant_layouts(&m.type_decls);
-        variant_layouts.by_type.extend(m_vl.by_type);
-        variant_layouts.ctor_to_type.extend(m_vl.ctor_to_type);
-        variant_layouts.ctor_field_defaults.extend(m_vl.ctor_field_defaults);
-    }
-    {
-        let mut owners: std::collections::HashMap<String, Vec<String>> = Default::default();
-        for k in variant_layouts.by_type.keys() {
-            if let Some((_, base)) = k.rsplit_once('.') {
-                owners.entry(base.to_string()).or_default().push(k.clone());
-            }
-        }
-        for (base, ks) in owners {
-            if ks.len() == 1 && !variant_layouts.by_type.contains_key(&base) {
-                let v = variant_layouts.by_type.get(&ks[0]).cloned().unwrap();
-                variant_layouts.by_type.insert(base, v);
-            }
-        }
-    }
-
-    PipelineLayouts {
-        globals,
-        global_inits,
-        main_globals,
-        main_global_inits,
-        mutable_toplet_aliases,
-        record_layouts,
-        variant_layouts,
-    }
-}
-
-/// The [`inline_and_classify_cross_module_fns`] outputs: the mangled linked-module
-/// sibling fns, MAIN's tail-inlined fns, and the union of both (the whole-program set
-/// the ABI-registry classification needs).
-struct CrossModuleFns {
-    module_fn_sibs: Vec<almide_ir::IrFunction>,
-    inlined_fns: Vec<almide_ir::IrFunction>,
-    all_fns: Vec<almide_ir::IrFunction>,
-}
-
-/// Phase 3: PROGRAM pre-pass — inline mutual-recursive tail siblings (semantics-preserving
-/// TCO exposure). The input is the WHOLE program — main's functions PLUS every linked
-/// user-module sibling under its MANGLED `almide_rt_<m>_<f>` name (bodies already
-/// reference siblings by that name, post-`resolve_user_module_calls`). Without the
-/// siblings, the never-err/auto-wrap ABI registries were populated from MAIN's functions
-/// only: a cross-module effect callee (`m.estep`) was UNCLASSIFIED, so the caller kept
-/// its auto-`?` Try (expecting a heap Result handle) while the separately-lowered callee
-/// returned its raw scalar — the crossmod_shape_matrix i64/i32 invalid-wasm class. One
-/// combined classification makes caller and callee agree by construction; the returned
-/// rewritten bodies are then split back into the main / module lowering regions (each
-/// keeps its own globals union) — iterated to a FIXPOINT (the #485 effect_assign shape).
-fn inline_and_classify_cross_module_fns(
-    ir: &almide_ir::IrProgram,
-    main_globals: &HashMap<almide_ir::VarId, almide_lang::types::Ty>,
-    record_layouts: &crate::lower::RecordLayouts,
-) -> CrossModuleFns {
-    let mut module_fn_sibs: Vec<almide_ir::IrFunction> = ir
-        .modules
-        .iter()
-        .filter(|m| !almide_lang::stdlib_info::is_any_stdlib(m.name.as_str()))
-        .flat_map(|m| {
-            let mname = m.name.as_str().to_string();
-            // INTRA-MODULE bare sibling calls resolve MODULE-LOCALLY (the #692 rule:
-            // current-module qualified > bare > any-module) — a clone body left with a
-            // bare `route(x, 100)` linked MAIN's same-named 0-arg `route` and shipped
-            // invalid wasm as "v1-verified" (values remaining on stack at the callee's
-            // arity mismatch — wasm_same_name_crossmod_test).
-            let sibs: std::collections::HashSet<String> =
-                m.functions.iter().map(|f| f.name.as_str().to_string()).collect();
-            m.functions.iter().filter(|f| !f.is_test).map(move |f| {
-                let mut nf = f.clone();
-                nf.name = almide_lang::intern::sym(&user_module_fn_name(&mname, f.name.as_str()));
-                struct Rw<'a> {
-                    mname: &'a str,
-                    sibs: &'a std::collections::HashSet<String>,
-                }
-                impl almide_ir::IrMutVisitor for Rw<'_> {
-                    fn visit_expr_mut(&mut self, e: &mut almide_ir::IrExpr) {
-                        almide_ir::walk_expr_mut(self, e);
-                        if let almide_ir::IrExprKind::Call { target, .. } = &mut e.kind {
-                            if let almide_ir::CallTarget::Named { name } = target {
-                                let f = name.as_str();
-                                if !f.starts_with("almide_rt_") && self.sibs.contains(f) {
-                                    *target = almide_ir::CallTarget::Named {
-                                        name: almide_lang::intern::sym(&user_module_fn_name(
-                                            self.mname, f,
-                                        )),
-                                    };
-                                }
-                            }
-                        }
-                    }
-                }
-                let mut rw = Rw { mname: &mname, sibs: &sibs };
-                almide_ir::IrMutVisitor::visit_expr_mut(&mut rw, &mut nf.body);
-                nf
-            })
-        })
-        .collect();
-    let mut all_fns: Vec<almide_ir::IrFunction> = ir.functions.clone();
-    all_fns.extend(module_fn_sibs.iter().cloned());
-    // The combined run POPULATES the name-keyed ABI registries over the WHOLE program
-    // (that is the crossmod fix — caller and callee classify identically); only MAIN's
-    // rewritten bodies are kept. The module siblings lower below from their ORIGINAL
-    // bodies: feeding the pre-pass's REWRITTEN module bodies through the module loop
-    // regressed the intra-module tail-call shape (`route(x, 100)` left values on the
-    // wasm stack — wasm_same_name_crossmod_test), and the registries alone are what the
-    // module-side lowering consults by MANGLED name.
-    let mut inlined_fns =
-        crate::lower::inline_mutual_tail_recursion(&ir.functions, main_globals, record_layouts);
-    // WIDEN the ABI registries over the whole program AFTER the main pre-pass (whose own
-    // population is main-only, the pre-batch behavior its rewrites were verified under):
-    // every LOWERING-time keyed lookup (never-err strip exclusions, AUTO_WRAP body.ty
-    // override, `ret_is_result_abi`) then sees module callees by their mangled names —
-    // the crossmod caller/callee ABI agreement — without the pre-pass rewrites ever
-    // touching module bodies.
-    crate::lower::populate_abi_registries(&all_fns, record_layouts);
-    // The registries above are program-wide, but the never-err REWRITES ran with MAIN-ONLY
-    // sets (inside the pre-pass) and never touched module bodies at all. So a MAIN caller of
-    // a cross-module never-err callee kept its lifted `Try`/Result-typed call, and a MODULE
-    // sibling caller kept its own — both then lower Result-handle reads (`local.set`) over
-    // the raw/void ABI the callee's def actually has: the #786 invalid-wasm class. Re-run
-    // the never-err rewrite family with the PROGRAM-WIDE sets over BOTH regions, so every
-    // caller agrees with the combined classification by construction (idempotent where the
-    // main-only pass already fired; call TARGETS are never renamed, so the module-side name
-    // resolution the original-bodies rule protects is untouched).
-    //
-    // FIXPOINT (the #485 effect_assign shape): the strips consult AUTO_WRAP (an
-    // auto-wrapped callee's `!`/`??` must NOT strip) while AUTO_WRAP itself is derived
-    // from the bodies (has a stmt-position propagating unwrap). A strip can remove a
-    // callee's LAST propagating unwrap (`plain_assign`'s `x = step(x)` Try), after which
-    // its REAL lowered ABI is bare — but the stale registry still said "wrapped", so its
-    // own def lowered bare while every `plain_assign()!` site kept the Result-handle
-    // read: invalid wasm (def/callsite ABI split). Iterate populate → rewrite until the
-    // registries describe the rewritten bodies verbatim (monotone — strips only remove
-    // nodes, AUTO_WRAP only shrinks — so this terminates; the cap is a safety net).
-    let mut prev_auto_wrap: Option<std::collections::HashSet<String>> = None;
-    for _ in 0..8 {
-        let mut all_rewritten: Vec<almide_ir::IrFunction> = inlined_fns.clone();
-        all_rewritten.extend(module_fn_sibs.iter().cloned());
-        crate::lower::populate_abi_registries(&all_rewritten, record_layouts);
-        let cur = crate::lower::auto_wrap_abi_snapshot();
-        if prev_auto_wrap.as_ref() == Some(&cur) {
-            break;
-        }
-        prev_auto_wrap = Some(cur);
-        let wide_can_err = crate::lower::compute_can_err(&all_rewritten);
-        let wide_lifted = crate::lower::lifted_effect_fn_names(&all_rewritten);
-        for f in inlined_fns.iter_mut().chain(module_fn_sibs.iter_mut()) {
-            let self_name = f.name.as_str().to_string();
-            crate::lower::strip_never_err_unwraps(
-                &mut f.body,
-                &wide_can_err,
-                &wide_lifted,
-                &self_name,
-            );
-            crate::lower::rewrite_never_err_effect_match(&mut f.body, &wide_can_err, &wide_lifted);
-            crate::lower::unwrap_never_err_call_types(&mut f.body, &wide_can_err, &wide_lifted);
-            crate::lower::rewrap_never_err_into_result_targets(
-                &mut f.body,
-                &wide_can_err,
-                &wide_lifted,
-                record_layouts,
-            );
-        }
-    }
-    CrossModuleFns { module_fn_sibs, inlined_fns, all_fns }
-}
-
-/// Phase 4: cross-module DERIVED-METHOD name bridge (#790 codec row, piece 2 of the
-/// pinned design): a MAIN-region `T.encode` / `T.decode` reference whose type `T` is
-/// declared by exactly ONE linked module (and not by main) resolves to that module's
-/// MANGLED derived fn (`almide_rt_<m>_T.encode`) — the same unique-owner rule the
-/// variant-layout bridging above uses. Without this the reference stays unlinked and
-/// the whole program walls (honest, but the direct-method shapes are fully lowerable).
-/// Container helpers (`__encode_list_<m>.T`) stay walled — their v1 lowering is the
-/// recorded remainder of the bridge design. Rewrites `inlined_fns`/`module_fn_sibs`
-/// bodies in place (both regions: main's derived fns reference the imported payload
-/// type's codec methods, and the OWNING module's own derived fns reference their
-/// sibling types' methods by the same bare `T.method` names).
-fn bridge_cross_module_derived_methods(
-    ir: &almide_ir::IrProgram,
-    inlined_fns: &mut [almide_ir::IrFunction],
-    module_fn_sibs: &mut [almide_ir::IrFunction],
-) {
-    let main_types: std::collections::HashSet<&str> =
-        ir.type_decls.iter().map(|td| td.name.as_str()).collect();
-    let mut owners: std::collections::HashMap<&str, Vec<&str>> = std::collections::HashMap::new();
-    for m in &ir.modules {
-        if almide_lang::stdlib_info::is_any_stdlib(m.name.as_str()) {
-            continue;
-        }
-        for td in &m.type_decls {
-            // Module type names may arrive QUALIFIED (`varlib.Pigment`) — key the
-            // owner map by the BASE name (the same normalization the variant-layout
-            // bridging above applies).
-            let base = td.name.as_str().rsplit('.').next().unwrap_or(td.name.as_str());
-            owners.entry(base).or_default().push(m.name.as_str());
-        }
-    }
-    struct Rw<'a> {
-        main_types: &'a std::collections::HashSet<&'a str>,
-        owners: &'a std::collections::HashMap<&'a str, Vec<&'a str>>,
-    }
-    impl almide_ir::IrMutVisitor for Rw<'_> {
-        fn visit_expr_mut(&mut self, e: &mut almide_ir::IrExpr) {
-            almide_ir::walk_expr_mut(self, e);
-            if let almide_ir::IrExprKind::Call {
-                target: almide_ir::CallTarget::Named { name },
-                ..
-            } = &mut e.kind
-            {
-                let n = name.as_str();
-                if n.starts_with("almide_rt_") || n.starts_with("__") {
-                    return;
-                }
-                let Some((ty_name, method)) = n.rsplit_once('.') else { return };
-                if method != "encode" && method != "decode" {
-                    return;
-                }
-                // `varlib.Pigment.decode` → qualifier "varlib" + base "Pigment";
-                // `Pigment.decode` → base only. A qualified ref must match the
-                // owner; a bare ref must not shadow a MAIN type of the same name.
-                let (qualifier, base) = match ty_name.rsplit_once('.') {
-                    Some((q, b)) => (Some(q), b),
-                    None => (None, ty_name),
-                };
-                if qualifier.is_none() && self.main_types.contains(base) {
-                    return;
-                }
-                if let Some(ms) = self.owners.get(base) {
-                    if let [only] = ms.as_slice() {
-                        if qualifier.is_none() || qualifier == Some(only) {
-                            *name = almide_lang::intern::sym(&user_module_fn_name(
-                                only,
-                                &format!("{base}.{method}"),
-                            ));
-                        }
-                    }
-                }
-            }
-        }
-    }
-    let mut rw = Rw { main_types: &main_types, owners: &owners };
-    for f in inlined_fns.iter_mut().chain(module_fn_sibs.iter_mut()) {
-        almide_ir::IrMutVisitor::visit_expr_mut(&mut rw, &mut f.body);
-    }
-    // …and publish the unique-owner map for the DESUGAR-time resolution: the
-    // `T.method` Named names are FORMED inside the per-fn lowering (from Method
-    // targets), after this pipeline pass — the registry is how they see it.
-    let derived_owners: std::collections::HashMap<String, String> = owners
-        .iter()
-        .filter(|(t, ms)| ms.len() == 1 && !main_types.contains(*t))
-        .map(|(t, ms)| (t.to_string(), ms[0].to_string()))
-        .collect();
-    crate::lower::set_derived_type_owners(derived_owners);
-}
-
-/// Phase 5: MUTABLE module-level `var`s (program + modules) — assign each a
-/// linear-memory storage slot (declaration order = VarId order, the same ordering
-/// `__global_init` uses) and publish the VarId → (slot, Ty) map — reads/assigns then
-/// route through the slot (`Load`/`$__mg_get`/`$__mg_take`+`Store`). A VarId collision
-/// across regions or an over-cap count WALLS the program (honest, never a mis-routed
-/// slot). Returns the sorted mutable top-lets (their declaration order IS the slot
-/// order, needed again by `__mg_init` synthesis below).
-fn assign_mutable_global_slots(
-    ir: &almide_ir::IrProgram,
-    mutable_toplet_aliases: &std::collections::HashMap<almide_ir::VarId, almide_ir::VarId>,
-) -> Result<Vec<almide_ir::IrTopLet>, LowerError> {
-    let mut mutable_tls: Vec<_> = ir
-        .top_lets
-        .iter()
-        .chain(ir.modules.iter().flat_map(|m| m.top_lets.iter()))
-        .filter(|tl| tl.mutable)
-        .cloned()
-        .collect();
-    mutable_tls.sort_by_key(|tl| tl.var.0);
-    if mutable_tls.len() > 64 {
-        return Err(LowerError::Unsupported(format!(
-            "{} mutable module-level vars exceed the 64-slot global region",
-            mutable_tls.len()
-        )));
-    }
-    {
-        let mut seen = std::collections::HashSet::new();
-        for tl in &mutable_tls {
-            if !seen.insert(tl.var.0) {
-                return Err(LowerError::Unsupported(format!(
-                    "mutable module-level var id collision across regions ({:?})",
-                    tl.var
-                )));
-            }
-        }
-    }
-    let mut mutable_global_map: std::collections::HashMap<u32, (u32, almide_lang::types::Ty)> = mutable_tls
-        .iter()
-        .enumerate()
-        .map(|(i, tl)| (tl.var.0, (i as u32, tl.ty.clone())))
-        .collect();
-    // #782: alias each main-side synthesized ref onto its module var's slot —
-    // the retired v0 fallback used to absorb these as walls; now `m.count`
-    // reads and assigns route through the SAME storage the owning module uses.
-    for (main_id, mod_id) in mutable_toplet_aliases {
-        if let Some(entry) = mutable_global_map.get(&mod_id.0).cloned() {
-            mutable_global_map.insert(main_id.0, entry);
-        }
-    }
-    crate::lower::set_mutable_global_vars(mutable_global_map);
-    Ok(mutable_tls)
-}
-
-/// Phase 6: repair CROSS-MODULE global refs whose expr type the frontend never
-/// inferred (`v.white` — the ceangal theme class) from the bridged globals maps
-/// BEFORE lowering (or the AllTypesConcrete precondition walls the whole fn), then
-/// SUBSTITUTE every BRIDGED cross-module ref whose module-side init is a pure call
-/// or heap ctor (`v.black` → view's `let black = rgb(0,0,0)`) into the referencing fn
-/// bodies — a lowering-time CallFn there would break the classify `mir == ir` count,
-/// so the call must exist in the IR itself instead (the `inline_pure_call_globals`
-/// discipline, extended across the cross-module bridge). Purity gate: every call
-/// inside the init is a pure stdlib call or a RESOLVED user fn that is itself
-/// call-transitively clean (effect fns were mangled through the resolver and appear
-/// in no pure set). MAIN-region readers take the BIND form (a fn-top `let __g_init =
-/// …` plus a `Var` reference — a call spliced into an arbitrary position, e.g. a
-/// record field, would render as an invalid dst-less bare call); module siblings keep
-/// the raw substitution (their separate VarId region cannot carry a main-region bind
-/// id).
-fn repair_and_substitute_globals(
-    ir: &mut almide_ir::IrProgram,
-    inlined_fns: &mut [almide_ir::IrFunction],
-    module_fn_sibs: &mut [almide_ir::IrFunction],
-    layouts: &PipelineLayouts,
-    all_fns: &[almide_ir::IrFunction],
-) {
-    for f in inlined_fns.iter_mut() {
-        crate::lower::repair_unknown_global_ref_tys(f, &layouts.main_globals);
-        crate::lower::repair_member_field_tys(f, &layouts.record_layouts);
-    }
-    for f in module_fn_sibs.iter_mut() {
-        crate::lower::repair_unknown_global_ref_tys(f, &layouts.globals);
-        crate::lower::repair_member_field_tys(f, &layouts.record_layouts);
-    }
-
-    use almide_ir::visit::{walk_expr, IrVisitor};
-    struct HasImpure<'a> {
-        impure: bool,
-        effectish: &'a std::collections::HashSet<String>,
-    }
-    impl IrVisitor for HasImpure<'_> {
-        fn visit_expr(&mut self, e: &almide_ir::IrExpr) {
-            match &e.kind {
-                almide_ir::IrExprKind::RuntimeCall { .. } => self.impure = true,
-                almide_ir::IrExprKind::Call { target, .. } => match target {
-                    almide_ir::CallTarget::Module { module, func, .. } => {
-                        if !crate::purity::is_pure(module.as_str(), func.as_str()) {
-                            self.impure = true;
-                        }
-                    }
-                    almide_ir::CallTarget::Named { name } => {
-                        if self.effectish.contains(name.as_str()) {
-                            self.impure = true;
-                        }
-                    }
-                    _ => self.impure = true,
-                },
-                _ => {}
-            }
-            walk_expr(self, e);
-        }
-    }
-    let effectish: std::collections::HashSet<String> = all_fns
-        .iter()
-        .filter(|f| f.is_effect)
-        .map(|f| f.name.as_str().to_string())
-        .collect();
-    let mut subs: Vec<(almide_ir::VarId, almide_ir::IrExpr)> = Vec::new();
-    for (i, info) in ir.var_table.entries.iter().enumerate() {
-        if info.module_origin.is_none() {
-            continue;
-        }
-        let id = almide_ir::VarId(i as u32);
-        let Some(init) = layouts.main_global_inits.get(&id) else { continue };
-        // #782: with the v0 fallback retired, a HEAP toplet whose init is a
-        // CTOR form (tuple/record/variant/some/ok — `let PAIR = ("a", 1)`,
-        // `let MOOD = Happy`) must ALSO substitute: value_or_global's CONST
-        // path only materializes flat literals (LitStr / all-literal List),
-        // and the old "computed init" wall used to fall back to v0. The
-        // bind-form substitution evaluates the pure ctor at fn-top — same
-        // discipline as the pure-call inits below.
-        let heap_ctor_init = crate::lower::is_heap_ty(&init.ty)
-            && !matches!(
-                &init.kind,
-                almide_ir::IrExprKind::LitStr { .. } | almide_ir::IrExprKind::List { .. }
-            );
-        if !crate::lower::expr_contains_call(init) && !heap_ctor_init {
-            continue;
-        }
-        let mut h = HasImpure { impure: false, effectish: &effectish };
-        h.visit_expr(init);
-        if !h.impure {
-            subs.push((id, init.clone()));
-        }
-    }
-    for (id, init) in &subs {
-        // MAIN-region readers take the BIND form instead of the raw expression
-        // substitution: `let __g_init = default_gap(); …Var(__g_init)…` — a call
-        // spliced into an arbitrary position (a record FIELD — the #785 shape)
-        // rendered as a dst-less bare call (invalid wasm), while a call in BIND
-        // position plus a Var reference is the proven single-file form. The bind
-        // goes at the fn-body top (the init is pure, so hoisting its evaluation
-        // is unobservable); fns that never reference the global are untouched.
-        for f in inlined_fns.iter_mut() {
-            fn references(e: &almide_ir::IrExpr, id: almide_ir::VarId) -> bool {
-                use almide_ir::visit::{walk_expr, IrVisitor};
-                struct V(almide_ir::VarId, bool);
-                impl IrVisitor for V {
-                    fn visit_expr(&mut self, e: &almide_ir::IrExpr) {
-                        if matches!(&e.kind, almide_ir::IrExprKind::Var { id } if *id == self.0)
-                        {
-                            self.1 = true;
-                        }
-                        walk_expr(self, e);
-                    }
-                }
-                let mut v = V(id, false);
-                v.visit_expr(e);
-                v.1
-            }
-            if !references(&f.body, *id) {
-                continue;
-            }
-            let nv = ir.var_table.alloc(
-                almide_lang::intern::sym("__g_init"),
-                init.ty.clone(),
-                almide_ir::Mutability::Let,
-                None,
-            );
-            let nv_ref = almide_ir::IrExpr {
-                kind: almide_ir::IrExprKind::Var { id: nv },
-                ty: init.ty.clone(),
-                span: None,
-                def_id: None,
-            };
-            f.body = almide_ir::substitute::substitute_var_in_expr(&f.body, *id, &nv_ref);
-            let bind_stmt = almide_ir::IrStmt {
-                kind: almide_ir::IrStmtKind::Bind {
-                    var: nv,
-                    mutability: almide_ir::Mutability::Let,
-                    ty: init.ty.clone(),
-                    value: init.clone(),
-                },
-                span: None,
-            };
-            if let almide_ir::IrExprKind::Block { stmts, .. } = &mut f.body.kind {
-                stmts.insert(0, bind_stmt);
-            } else {
-                // An EXPRESSION-form body (`effect fn main() -> Unit =
-                // println(m.CFG.name)`) has no statement list — wrap it in a
-                // Block so the fn-top bind exists. Without this the
-                // substitution left `Var(__g_init)` UNBOUND (#782, the
-                // record/variant toplet matrix cells).
-                let old_ty = f.body.ty.clone();
-                let old_span = f.body.span.clone();
-                let old = std::mem::replace(
-                    &mut f.body,
-                    almide_ir::IrExpr {
-                        kind: almide_ir::IrExprKind::Unit,
-                        ty: almide_lang::types::Ty::Unit,
-                        span: None,
-                        def_id: None,
-                    },
-                );
-                f.body = almide_ir::IrExpr {
-                    kind: almide_ir::IrExprKind::Block {
-                        stmts: vec![bind_stmt],
-                        expr: Some(Box::new(old)),
-                    },
-                    ty: old_ty,
-                    span: old_span,
-                    def_id: None,
-                };
-            }
-        }
-        // Module siblings keep the raw substitution (their separate VarId
-        // numbering region cannot carry a main-region bind id) — the ceangal
-        // in-module reader class this path has always served.
-        for f in module_fn_sibs.iter_mut() {
-            f.body = almide_ir::substitute::substitute_var_in_expr(&f.body, *id, init);
-        }
-    }
-    if !subs.is_empty() {
-        for f in inlined_fns.iter_mut() {
-            crate::lower::repair_record_literal_field_tys(f);
-        }
-    }
-}
-
-/// Phase 7: lower every non-test MAIN fn to MIR (a fn that walls is silently skipped,
-/// listed to stderr under `verbose`), then lower every linked USER-module sibling the
-/// target's resolved `almide_rt_<m>_<f>` references — under the SAME mangled name, so
-/// every keyed lookup (never-err strip, AUTO_WRAP ABI, `ret_is_result_abi`) sees what
-/// callers use via the combined registry population. Each module fn lowers SEPARATELY
-/// (its own VarId region + shared globals); one already defined (from `inlined_fns`,
-/// the main-region tail-inlining pass) or one that itself walls is silently skipped
-/// (the caller then fails the unlinked-call render wall if it truly needed it — stdlib
-/// modules stay out, self-host-linked below).
-fn lower_main_and_sibling_fns(
-    inlined_fns: &[almide_ir::IrFunction],
-    module_fn_sibs: &[almide_ir::IrFunction],
-    layouts: &PipelineLayouts,
-    total_ir_fn_count: usize,
-    verbose: bool,
-) -> Vec<crate::MirFunction> {
-    let mut functions = Vec::new();
-    let mut walled = Vec::new();
-    for func in inlined_fns {
-        // `test "…"` blocks lower to fns calling the test harness (no wasm def) — never reachable
-        // from `_start`/`main`, so skip them (rendering one would pull a dangling `(call $assert_eq)`).
-        if func.is_test {
-            continue;
-        }
-        match crate::lower::lower_function_all_with_globals(
-            func,
-            &layouts.main_globals,
-            &layouts.main_global_inits,
-            &layouts.record_layouts,
-            &layouts.variant_layouts,
-        ) {
-            Ok(mirs) => functions.extend(mirs),
-            Err(e) => walled.push(format!("{}: {e:?}", func.name.as_str())),
-        }
-    }
-    if !walled.is_empty() && verbose {
-        eprintln!(
-            "[render_program] {} of {} function(s) outside the lowering subset (NOT rendered):",
-            walled.len(),
-            total_ir_fn_count
-        );
-        for w in &walled {
-            eprintln!("  {w}");
-        }
-    }
-
-    let already: std::collections::HashSet<String> =
-        functions.iter().map(|f| f.name.clone()).collect();
-    for func in module_fn_sibs {
-        // ORIGINAL bodies under the mangled name — every keyed lookup (never-err strip,
-        // AUTO_WRAP ABI, `ret_is_result_abi`) sees the SAME name callers use via the
-        // combined registry population above.
-        if already.contains(func.name.as_str()) {
-            continue;
-        }
-        if let Ok(mirs) = crate::lower::lower_function_all_with_globals(
-            func,
-            &layouts.globals,
-            &layouts.global_inits,
-            &layouts.record_layouts,
-            &layouts.variant_layouts,
-        ) {
-            functions.extend(mirs);
-        }
-    }
-    functions
-}
-
-/// Phase 8: synthesize `__mg_init` (assigns each mutable top-let its declared
-/// initializer, declaration order, through the SAME slot-routed Assign path user code
-/// uses — `_start` calls it before `__global_init`/`main`; a non-lowerable initializer
-/// WALLS the whole program, since shipping zeroed globals would be a silent
-/// miscompile), then auto-link the self-hosted stdlib runtime (int.to_string,
-/// string.concat, `print_str`, …): when a registry entry is called but not defined,
-/// its impl fn is lowered and renamed to the call name — iterated to a FIXPOINT since
-/// a linked impl may itself call ANOTHER registry entry. A self-hosted fn calling
-/// another registered impl by its IMPL name (pre-rename) is rewritten to the call name
-/// afterward, and `print_str` is force-linked last (`println` → `PrintStr` → `(call
-/// $print_str)`).
-#[allow(clippy::too_many_arguments)]
-fn synthesize_and_link_runtime_fns(
-    functions: &mut Vec<crate::MirFunction>,
-    mutable_tls: &[almide_ir::IrTopLet],
-    layouts: &PipelineLayouts,
-    verbose: bool,
-) -> Result<(), LowerError> {
-    if !mutable_tls.is_empty() {
-        let stmts: Vec<almide_ir::IrStmt> = mutable_tls
-            .iter()
-            .map(|tl| almide_ir::IrStmt {
-                kind: almide_ir::IrStmtKind::Assign { var: tl.var, value: tl.value.clone() },
-                span: None,
-            })
-            .collect();
-        let body = almide_ir::IrExpr {
-            kind: almide_ir::IrExprKind::Block { stmts, expr: None },
-            ty: almide_lang::types::Ty::Unit,
-            span: Default::default(),
-            def_id: None,
-        };
-        let init_fn = almide_ir::IrFunction {
-            name: almide_lang::intern::sym("__mg_init"),
-            params: vec![],
-            ret_ty: almide_lang::types::Ty::Unit,
-            body,
-            is_effect: false,
-            is_async: false,
-            is_test: false,
-            generics: None,
-            extern_attrs: vec![],
-            export_attrs: vec![],
-            attrs: vec![],
-            visibility: almide_ir::IrVisibility::Public,
-            doc: None,
-            blank_lines_before: 0,
-            def_id: None,
-            module_origin: None,
-            mutated_params: vec![],
-        };
-        match crate::lower::lower_function_all_with_globals(
-            &init_fn,
-            &layouts.main_globals,
-            &layouts.main_global_inits,
-            &layouts.record_layouts,
-            &layouts.variant_layouts,
-        ) {
-            Ok(mirs) => functions.extend(mirs),
-            Err(e) => {
-                return Err(LowerError::Unsupported(format!(
-                    "mutable module-level var initializer outside the executable subset: {e:?}"
-                )))
-            }
-        }
-    }
-
-    // Auto-link the self-hosted stdlib runtime (int.to_string, string.concat, …) when an entry is
-    // called but not defined, renaming its impl fn to the call name. A linked impl may call ANOTHER
-    // registry entry, so iterate to a FIXPOINT.
-    loop {
-        let before = functions.len();
-        for (rt_source, entries) in crate::render_wasm::self_host_runtime() {
-            let mut any_called = entries.iter().any(|(_, call)| {
-                functions.iter().any(|f| {
-                    f.ops.iter().any(|op| matches!(op, crate::Op::CallFn { name, .. } if name == call))
-                })
-            });
-            // A Value drop renders `(call $__drop_value …)` — a value_core helper that is NOT a
-            // registered call_name, so force value_core when ANY Value-drop op is present.
-            if entries.iter().any(|(_, c)| *c == "value.null") {
-                any_called = any_called
-                    || functions.iter().any(|f| {
-                        f.ops.iter().any(|op| {
-                            matches!(
-                                op,
-                                crate::Op::DropValue { .. }
-                                    | crate::Op::DropListValue { .. }
-                                    | crate::Op::DropListStrValue { .. }
-                                    | crate::Op::DropListStrStr { .. }
-                                    | crate::Op::DropResultValue { .. }
-                                    | crate::Op::DropResultListValue { .. }
-                            )
-                        })
-                    });
-            }
-            // A `Map[String, <flat heap>]` drop renders `(call $__drop_map_hval …)` — a
-            // map_hval helper that is NOT a registered call name, so force map_hval when
-            // the DropVariant is present (the C-039 typechange twins build hval-flavor
-            // maps without calling any registered map_hval fn — the value_core pattern).
-            if entries.iter().any(|(impl_fn, _)| *impl_fn == "map_new_hval") {
-                any_called = any_called
-                    || functions.iter().any(|f| {
-                        f.ops.iter().any(|op| matches!(op,
-                            crate::Op::DropVariant { ty, .. } if ty == "map_hval"))
-                    });
-            }
-            let any_defined =
-                entries.iter().any(|(_, call)| functions.iter().any(|f| &f.name == call));
-            if any_called && !any_defined {
-                let rt = source_to_ir(rt_source)?;
-                for f in &rt.functions {
-                    let lowered = crate::lower::lower_function(f, &layouts.globals);
-                    if let Err(e) = &lowered {
-                        if verbose
-                            && (entries.iter().any(|(impl_fn, _)| f.name.as_str() == *impl_fn)
-                                || f.name.as_str().starts_with("__"))
-                        {
-                            eprintln!("[self-host] {} failed to lower: {:?}", f.name.as_str(), e);
-                        }
-                    }
-                    if let Ok(mut mir) = lowered {
-                        if let Some((_, call)) =
-                            entries.iter().find(|(impl_fn, _)| &mir.name == impl_fn)
-                        {
-                            mir.name = call.to_string();
-                        }
-                        functions.push(mir);
-                    }
-                }
-            }
-        }
-        // Dedup by name (identical source ⇒ no-op merge).
-        let mut seen = std::collections::HashSet::new();
-        functions.retain(|f| seen.insert(f.name.clone()));
-        if functions.len() == before {
-            break;
-        }
-    }
-
-    // A self-hosted runtime fn may call ANOTHER registered impl by its IMPL name, but the auto-link
-    // RENAMED that def to its call_name. Rewrite those call sites to the call_name.
-    let impl_to_call: std::collections::HashMap<&str, &str> = crate::render_wasm::self_host_runtime()
-        .iter()
-        .flat_map(|(_, es)| es.iter().map(|(i, c)| (*i, *c)))
-        .collect();
-    for f in functions.iter_mut() {
-        for op in &mut f.ops {
-            if let crate::Op::CallFn { name, .. } = op {
-                if let Some(&c) = impl_to_call.get(name.as_str()) {
-                    *name = c.to_string();
-                }
-            }
-        }
-    }
-
-    // Auto-link the self-hosted runtime `print_str` (`println` → `PrintStr` → `(call $print_str)`).
-    if !functions.iter().any(|f| f.name == "print_str") {
-        let rt = source_to_ir(include_str!("../../../stdlib/print_str.almd"))?;
-        for f in &rt.functions {
-            if let Ok(mir) = crate::lower::lower_function(f, &layouts.globals) {
-                functions.push(mir);
-            }
-        }
-    }
-    Ok(())
-}
-
-fn try_render_wasm_source_impl_rest(
-    ir: &mut almide_ir::IrProgram,
-    verbose: bool,
-) -> Result<String, LowerError> {
-    let layouts = collect_pipeline_layouts(ir);
-
-    let CrossModuleFns { mut module_fn_sibs, mut inlined_fns, all_fns } =
-        inline_and_classify_cross_module_fns(ir, &layouts.main_globals, &layouts.record_layouts);
-
-    bridge_cross_module_derived_methods(ir, &mut inlined_fns, &mut module_fn_sibs);
-
-    let mutable_tls = assign_mutable_global_slots(ir, &layouts.mutable_toplet_aliases)?;
-    let mutable_global_count = mutable_tls.len() as u32;
-
-    repair_and_substitute_globals(ir, &mut inlined_fns, &mut module_fn_sibs, &layouts, &all_fns);
-
-    let mut functions = lower_main_and_sibling_fns(
-        &inlined_fns,
-        &module_fn_sibs,
-        &layouts,
-        ir.functions.len(),
-        verbose,
-    );
-
-    synthesize_and_link_runtime_fns(&mut functions, &mutable_tls, &layouts, verbose)?;
-
-    // EAGER GLOBAL-INIT semantics (C-007): v0 evaluates every ABORTABLE top-let initializer at
-    // startup. Synthesize `__global_init` binding each CALL-FREE SCALAR initializer and have
-    // `_start` call it before `$main`. Call-bearing/heap inits keep per-use/wall handling.
-    {
-        fn has_call(e: &almide_ir::IrExpr) -> bool {
-            use almide_ir::visit::{walk_expr, IrVisitor};
-            struct C(bool);
-            impl IrVisitor for C {
-                fn visit_expr(&mut self, e: &almide_ir::IrExpr) {
-                    if matches!(
-                        e.kind,
-                        almide_ir::IrExprKind::Call { .. } | almide_ir::IrExprKind::RuntimeCall { .. }
-                    ) {
-                        self.0 = true;
-                    }
-                    walk_expr(self, e);
-                }
-            }
-            let mut c = C(false);
-            c.visit_expr(e);
-            c.0
-        }
-        let mut max_var = 0u32;
-        for (v, _) in &layouts.globals {
-            max_var = max_var.max(v.0);
-        }
-        {
-            use almide_ir::visit::{walk_expr, IrVisitor};
-            struct M(u32);
-            impl IrVisitor for M {
-                fn visit_expr(&mut self, e: &almide_ir::IrExpr) {
-                    if let almide_ir::IrExprKind::Var { id } = &e.kind {
-                        self.0 = self.0.max(id.0);
-                    }
-                    walk_expr(self, e);
-                }
-            }
-            let mut m = M(max_var);
-            for f in &ir.functions {
-                m.visit_expr(&f.body);
-            }
-            max_var = m.0;
-        }
-        let mut stmts: Vec<almide_ir::IrStmt> = Vec::new();
-        let mut ordered: Vec<_> = ir
-            .top_lets
-            .iter()
-            .chain(ir.modules.iter().flat_map(|m| m.top_lets.iter()))
-            .collect();
-        ordered.sort_by_key(|tl| tl.var.0);
-        // A later initializer may READ an earlier global — INLINE each processed init into its
-        // dependents (declaration order; all call-free ⇒ pure, finite substitution).
-        let mut subst: std::collections::HashMap<almide_ir::VarId, almide_ir::IrExpr> =
-            std::collections::HashMap::new();
-        for tl in ordered {
-            let scalar = !crate::lower::is_heap_ty(&tl.ty);
-            if scalar && !has_call(&tl.value) {
-                let mut value = tl.value.clone();
-                for (gv, ge) in &subst {
-                    value = almide_ir::substitute::substitute_var_in_expr(&value, *gv, ge);
-                }
-                subst.insert(tl.var, value.clone());
-                max_var += 1;
-                stmts.push(almide_ir::IrStmt {
-                    kind: almide_ir::IrStmtKind::Bind {
-                        var: almide_ir::VarId(max_var),
-                        mutability: almide_ir::Mutability::Let,
-                        ty: tl.ty.clone(),
-                        value,
-                    },
-                    span: None,
-                });
-            }
-        }
-        if !stmts.is_empty() {
-            let body = almide_ir::IrExpr {
-                kind: almide_ir::IrExprKind::Block { stmts, expr: None },
-                ty: almide_lang::types::Ty::Unit,
-                span: Default::default(),
-                def_id: None,
-            };
-            let init_fn = almide_ir::IrFunction {
-                name: almide_lang::intern::sym("__global_init"),
-                params: vec![],
-                ret_ty: almide_lang::types::Ty::Unit,
-                body,
-                is_effect: false,
-                is_async: false,
-                is_test: false,
-                generics: None,
-                extern_attrs: vec![],
-                export_attrs: vec![],
-                attrs: vec![],
-                visibility: almide_ir::IrVisibility::Public,
-                doc: None,
-                blank_lines_before: 0,
-                def_id: None,
-                module_origin: None,
-                mutated_params: vec![],
-            };
-            if let Ok(mir) = crate::lower::lower_function(&init_fn, &layouts.globals) {
-                functions.push(mir);
-            }
-        }
-    }
-
-    // If `main` itself was WALLED, there is no `$main` — yet the renderer emits
-    // `(func (export "_start") (call $main))`. Wall the WHOLE program cleanly instead of a
-    // main-less (invalid) module.
-    if !functions.iter().any(|f| f.name == "main") {
-        return Err(LowerError::Unsupported(
-            "main is outside the MIR-lowering subset".into(),
-        ));
-    }
-
-    // `pub fn` EXPORT roots (#457): a Public non-test MAIN-program fn must be a named wasm
-    // export (host-invocable, the v0 emitter's export contract). One that LOWERED gets an
-    // `(export …)` directive; one that WALLED cannot be exported — decline the WHOLE module
-    // so the `--verified` pipeline falls back to v0 (which exports it) rather than shipping
-    // an artifact silently missing a public entry point.
-    let mut exports: Vec<(String, String, Vec<bool>, Option<bool>)> = Vec::new();
-    let is_float_ty = |t: &almide_lang::types::Ty| {
-        matches!(
-            t,
-            almide_lang::types::Ty::Float
-                | almide_lang::types::Ty::Float32
-                | almide_lang::types::Ty::Float64
-        )
-    };
-    for func in &ir.functions {
-        if !func.is_test
-            && func.name.as_str() != "main"
-            && !func.generics.as_ref().map_or(false, |g| !g.is_empty())
-            && matches!(func.visibility, almide_ir::IrVisibility::Public)
-        {
-            let n = func.name.as_str();
-            if functions.iter().any(|f| f.name == n) {
-                // `@export(wasm, "sym")` overrides the export name (v0's criterion,
-                // mod_p3.rs). v1's internal value model carries a Float as raw i64 BITS;
-                // the renderer emits a reinterpret wrapper for Float-bearing signatures
-                // so the public ABI presents real f64s (v0 parity).
-                let export_name = func
-                    .export_attrs
-                    .iter()
-                    .find(|a| a.target.as_str() == "wasm")
-                    .map(|a| a.symbol.to_string())
-                    .unwrap_or_else(|| n.to_string());
-                let param_floats: Vec<bool> =
-                    func.params.iter().map(|p| is_float_ty(&p.ty)).collect();
-                let ret_float = match &func.ret_ty {
-                    almide_lang::types::Ty::Unit => None,
-                    t => Some(is_float_ty(t)),
-                };
-                exports.push((export_name, n.to_string(), param_floats, ret_float));
-            } else {
-                return Err(LowerError::Unsupported(format!(
-                    "exported `pub fn {n}` is outside the MIR-lowering subset (the wasm module \
-                     must carry its export)"
-                )));
-            }
-        }
-    }
-
-    // #824: drop MakeUnique guards that are provably dead (the value they'd guard
-    // is never aliased anywhere in its own function) — see alias_safety.rs's doc
-    // comment for the soundness argument. Target-agnostic (applies before either
-    // renderer runs), so the native leg gets the identical benefit below too.
-    crate::alias_safety::elide_unaliased_make_unique(&mut functions);
-
-    // Any UNLINKED stdlib/runtime call would render a dangling `(call $name)` (invalid wasm) — the
-    // renderer rejects it cleanly. Returns the WAT on success.
-    try_render_wasm_program(&MirProgram { functions, exports, mutable_global_count })
-}
-
-/// NATIVE leg of the trust spine (#764, rung 1): lower `.almd` source through the
-/// SAME Perceus MIR the wasm leg uses and render it to native Rust — `Dup` as
-/// `.clone()`, `Drop` erased to Rust's scope-end drop, the runtime boundary mapped
-/// to a closed native shim floor. `verify_ownership` certifies the Perceus balance
-/// on the same ops before the erasure. WALLS (`Err`) on anything outside the
-/// rung-1 subset — the CLI falls back to v0, so a rendered program is never wrong.
-/// Debug probe: dump the lowered MIR ops of every non-test fn (walls listed
-/// per fn). Used by examples/probe_native.rs during rung development.
-pub fn debug_dump_mir(source: &str) -> Result<String, LowerError> {
-    crate::lower::STRICT_VALUES.store(true, std::sync::atomic::Ordering::Relaxed);
-    let ir = source_to_ir_with(source, &[])?;
-    let globals = std::collections::HashMap::new();
-    let global_inits = std::collections::HashMap::new();
-    // The REAL pipeline's layout registries — without them a record literal
-    // lowers as an Opaque skeleton and a field read walls (the very first
-    // rung-5 records probe misread that as a lowering gap).
-    let record_layouts = crate::lower::build_record_layouts(&ir.type_decls);
-    let variant_layouts = crate::lower::build_variant_layouts(&ir.type_decls);
-    let mut out = String::new();
-    for func in &ir.functions {
-        if func.is_test {
-            continue;
-        }
-        match crate::lower::lower_function_all_with_globals(func, &globals, &global_inits, &record_layouts, &variant_layouts) {
-            Ok(all) => {
-                for f in all {
-                    out.push_str(&format!("== fn {} ==\n", f.name));
-                    for op in &f.ops {
-                        out.push_str(&format!("  {op:?}\n"));
-                    }
-                }
-            }
-            Err(e) => out.push_str(&format!("== fn {} LOWER-WALL: {e:?}\n", func.name)),
-        }
-    }
-    Ok(out)
-}
-
-pub fn try_render_rust_source(source: &str) -> Result<String, LowerError> {
-    crate::lower::STRICT_VALUES.store(true, std::sync::atomic::Ordering::Relaxed);
-    let ir = source_to_ir_with(source, &[])?;
-    // Rung-5 records slab: the layout registries the wasm leg threads — without
-    // them a record literal lowers as an Opaque skeleton and every field read
-    // strict-walls (the probe_native trap recorded in the trust-spine ledger).
-    let record_layouts = crate::lower::build_record_layouts(&ir.type_decls);
-    let variant_layouts = crate::lower::build_variant_layouts(&ir.type_decls);
-    if !ir.modules.is_empty() {
-        return Err(LowerError::Unsupported(
-            "native: multi-module program — outside rung 1".into(),
-        ));
-    }
-    if !ir.top_lets.is_empty() {
-        return Err(LowerError::Unsupported(
-            "native: top-level lets — outside rung 1".into(),
-        ));
-    }
-    let globals = std::collections::HashMap::new();
-    let mut functions = Vec::new();
-    for func in &ir.functions {
-        if func.is_test {
-            continue;
-        }
-        // PRECISION WALL (rung 2): the native renderer types a heap `Repr::Ptr`
-        // param/result as a STRING — only sound when the DECLARED type says so.
-        // Any signature outside {Int, Bool, String, Unit-ret} walls here, where
-        // the Almide-level `Ty` is still visible (`MirParam` carries only reprs).
-        use almide_lang::types::Ty;
-        use almide_lang::types::constructor::TypeConstructorId;
-        let is_scalar_list = |t: &Ty| {
-            matches!(t, Ty::Applied(TypeConstructorId::List, a)
-                if a.len() == 1 && matches!(a[0], Ty::Int | Ty::Bool))
-        };
-        // Rung-5 records slab: an ALL-SCALAR record is layout-identical to a
-        // scalar list (the DynList block), so its params/returns ride the same
-        // `&[i64]`/`Vec<i64>` convention on native.
-        let is_scalar_record = |t: &Ty| -> bool {
-            let Ty::Named(n, args) = t else { return false };
-            if !args.is_empty() { return false; }
-            record_layouts
-                .get(n.as_str())
-                .is_some_and(|(_, ftys)| ftys.iter().all(|(_, ft)| !crate::lower::is_heap_ty(ft)))
-        };
-        // A FLAT variant (every ctor scalar-only) is likewise one slot block
-        // (tag@0, payload@1+) — same `&[i64]` convention (rung-5 variants slab).
-        let is_flat_variant = |t: &Ty| -> bool {
-            let Ty::Named(n, args) = t else { return false };
-            if !args.is_empty() { return false; }
-            variant_layouts.by_type.get(n.as_str()).is_some_and(|layout| {
-                layout.cases.iter().all(|c| c.fields.iter().all(|(_, ft)| !crate::lower::is_heap_ty(ft)))
-            })
-        };
-        // Rung-5 closures slab: a SCALAR closure type (`(Int) -> Int` — every param
-        // and the return scalar) travels as its env block (`Vec<i64>`: [fnidx,
-        // drop-header, captures…]); invocation dispatches through the generated
-        // `__almd_ci_*` tables. Heap-param/-return closures stay wasm-only.
-        let is_scalar_fn = |t: &Ty| {
-            matches!(t, Ty::Fn { params, ret }
-                if params.iter().all(|p| matches!(p, Ty::Int | Ty::Bool))
-                    && matches!(**ret, Ty::Int | Ty::Bool))
-        };
-        let sig_ok = |t: &Ty| {
-            matches!(t, Ty::Int | Ty::Bool | Ty::Float | Ty::String)
-                || is_scalar_list(t)
-                || is_scalar_record(t)
-                || is_flat_variant(t)
-                || is_scalar_fn(t)
-        };
-        for p in &func.params {
-            if !sig_ok(&p.ty) {
-                return Err(LowerError::Unsupported(format!(
-                    "native: fn `{}` param `{:?}` type — outside the native rung subset",
-                    func.name, p.ty
-                )));
-            }
-        }
-        if !matches!(func.ret_ty, Ty::Unit) && !sig_ok(&func.ret_ty) {
-            return Err(LowerError::Unsupported(format!(
-                "native: fn `{}` return type {:?} — outside the native rung subset",
-                func.name, func.ret_ty
-            )));
-        }
-        // ALL-OR-NOTHING: any unlowerable fn walls the program (the native rungs
-        // have no per-fn fallback — a partial native binary cannot call into v0).
-        let all = crate::lower::lower_function_all_with_globals(
-            func,
-            &globals,
-            &std::collections::HashMap::new(),
-            &record_layouts,
-            &variant_layouts,
-        )
-        .map_err(|e| {
-            LowerError::Unsupported(format!("native: fn `{}`: {e:?}", func.name))
-        })?;
-        functions.extend(all);
-    }
-    if !functions.iter().any(|f| f.name == "main") {
-        return Err(LowerError::Unsupported(
-            "native: main is outside the MIR-lowering subset".into(),
-        ));
-    }
-    // The SIG-KIND table (rung 4): the declared param/return kinds, computed here
-    // where the Almide-level `Ty` is visible, so the native render can type a heap
-    // `Repr::Ptr` as `&str` vs `&[i64]` per the declaration.
-    let mut sigs: crate::render_native::NativeSigs = Default::default();
-    {
-        use almide_lang::types::constructor::TypeConstructorId;
-        use almide_lang::types::Ty;
-        use crate::render_native::NativeSigKind;
-        let kind = |t: &Ty| -> Option<NativeSigKind> {
-            match t {
-                Ty::Int | Ty::Bool => Some(NativeSigKind::I64),
-                Ty::Float => Some(NativeSigKind::F64),
-                Ty::String => Some(NativeSigKind::Str),
-                Ty::Applied(TypeConstructorId::List, a)
-                    if a.len() == 1 && matches!(a[0], Ty::Int | Ty::Bool) =>
-                {
-                    Some(NativeSigKind::ListI64)
-                }
-                // An all-scalar record travels as its slot block (see sig_ok).
-                Ty::Named(n, args)
-                    if args.is_empty()
-                        && record_layouts
-                            .get(n.as_str())
-                            .is_some_and(|(_, ftys)| ftys.iter().all(|(_, ft)| !crate::lower::is_heap_ty(ft))) =>
-                {
-                    Some(NativeSigKind::ListI64)
-                }
-                // A flat variant travels as its tag+payload slot block.
-                Ty::Named(n, args)
-                    if args.is_empty()
-                        && variant_layouts.by_type.get(n.as_str()).is_some_and(|layout| {
-                            layout.cases.iter().all(|c| c.fields.iter().all(|(_, ft)| !crate::lower::is_heap_ty(ft)))
-                        }) =>
-                {
-                    Some(NativeSigKind::ListI64)
-                }
-                // A scalar closure travels as its env block (rung-5 closures slab).
-                Ty::Fn { params, ret }
-                    if params.iter().all(|p| matches!(p, Ty::Int | Ty::Bool))
-                        && matches!(**ret, Ty::Int | Ty::Bool) =>
-                {
-                    Some(NativeSigKind::ListI64)
-                }
-                _ => None,
-            }
-        };
-        for func in &ir.functions {
-            if func.is_test {
-                continue;
-            }
-            let params: Option<Vec<_>> = func.params.iter().map(|p| kind(&p.ty)).collect();
-            let ret = if matches!(func.ret_ty, Ty::Unit) {
-                Some(None)
-            } else {
-                kind(&func.ret_ty).map(Some)
-            };
-            if let (Some(ps), Some(r)) = (params, ret) {
-                sigs.insert(func.name.as_str().to_string(), (ps, r));
-            }
-        }
-        // LIFTED lambdas exist only as MirFunctions (no IR sig): param 0 is the env
-        // block (`&[i64]`), the rest are the lambda's own params — SCALAR reprs only
-        // in this slab (a heap-param lambda walls the program: all-or-nothing, and
-        // its dispatch arm could not type). The return kind is body-derived by the
-        // renderer, so it is not declared here.
-        for f in &functions {
-            if !f.name.starts_with("__lambda_") {
-                continue;
-            }
-            let mut ps = vec![crate::render_native::NativeSigKind::ListI64];
-            for p in &f.params[1..] {
-                match p.repr {
-                    crate::Repr::Scalar { .. } => {
-                        ps.push(crate::render_native::NativeSigKind::I64)
-                    }
-                    _ => {
-                        return Err(LowerError::Unsupported(format!(
-                            "native: lambda `{}` heap param — outside the closures slab",
-                            f.name
-                        )))
-                    }
-                }
-            }
-            sigs.insert(f.name.clone(), (ps, Some(crate::render_native::NativeSigKind::I64)));
-        }
-    }
-    // #824: see the wasm leg's call above — `Op::MakeUnique` already renders to
-    // nothing on native (render_native.rs's `Op::Consume | Op::Borrow |
-    // Op::MakeUnique => {}`), so this is a no-op cleanup here, kept only so both
-    // legs run the identical target-agnostic MIR pass list.
-    crate::alias_safety::elide_unaliased_make_unique(&mut functions);
-    crate::render_native::try_render_native_program(
-        &MirProgram {
-            functions,
-            exports: Vec::new(),
-            // Rung 1 walls every top-let above, so there are no mutable-global slots.
-            mutable_global_count: 0,
-        },
-        &sigs,
-    )
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    // `Parser::parse()` is a recovery parser: an unparseable top-level bare
-    // statement (never valid Almide grammar — only fn/effect fn/type/let/
-    // var/trait/impl/test are top-level declarations) is dropped via
-    // `skip_to_next_decl()`, but `parse()` still returns `Ok` as long as some
-    // decl survived. Before this fix, `source_to_ir_with` only checked the
-    // `Result` and never inspected `Parser::errors`, so a source like this
-    // would silently compile with the bare `println` call missing from the
-    // output — a wall was expected instead.
-    #[test]
-    fn dropped_top_level_statement_walls_instead_of_silently_compiling() {
-        let source = r#"
-let x = 1
-let y = 2
-println("top-level-statement-should-not-be-silently-dropped")
-fn main() -> Unit = {
-  println("from-main")
-}
-"#;
-        let result = try_render_wasm_source(source, &[], false);
-        assert!(
-            result.is_err(),
-            "a source with a dropped top-level statement must wall, not silently compile"
-        );
-    }
-}
+include!("pipeline_b.rs");
+include!("pipeline_c.rs");
